restructure google resource to have its own chart and parser for data
